refactor: Create toDiscordHTML util
toDiscordHTML(string): string
- Returns the html string of markdown text containing emojis, using Discord's formatting rules
- Used with dangerouslySetInnerHTML
