+++ src/common/components/ContentObject.css
@@ -34,6 +34,7 @@
     vertical-align: middle;
     font-size: medium;
     font-family: sans-serif;
+    font-weight: bold;
     color: #FFFFFF;
 }
